+++ database/migrations/20250826164924_change_urls_to_text_upload_table.php
@@ -1,0 +1,29 @@
+<?php
+
+use Movim\Migration;
+
+use App\Contact;
+use App\Presence;
+use App\Roster;
+use App\Conference;
+
+use Illuminate\Database\Schema\Blueprint;
+
+class ChangeUrlsToTextUploadTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('upload', function (Blueprint $table) {
+            $table->text('geturl')->nullable()->change();
+            $table->text('puturl')->nullable()->change();
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('upload', function (Blueprint $table) {
+            $table->string('geturl')->nullable()->change();
+            $table->string('puturl')->nullable()->change();
+        });
+    }
+}